refactor(paykit): better folder structure that reflects medusa providers
