+++ apps/medusa-be/src/modules/payment-paykit/__tests__/gopay.unit.spec.ts
@@ -1,8 +1,8 @@
 import { PaymentActions } from "@medusajs/framework/utils"
 import { describe, expect, it, vi } from "vitest"
 import { PAYKIT_GOPAY_WEBHOOK_PATH } from "../config"
-import { PaykitGopayPaymentProvider } from "../gopay"
 import { getGopayProviderOptions } from "../runtime"
+import { PaykitGopayPaymentProvider } from "../services/gopay"
 import { createMockContainer, createMockPaykitClient } from "./helpers"
 
 describe("PaykitGopayPaymentProvider", () => {
